Fix building.
