+++ intuition.cabal
@@ -1,64 +1,34 @@
 cabal-version:      3.0
--- The cabal-version field refers to the version of the .cabal specification,
--- and can be different from the cabal-install (the tool) version and the
--- Cabal (the library) version you are using. As such, the Cabal (the library)
--- version used must be equal or greater than the version stated in this field.
--- Starting from the specification version 2.2, the cabal-version field must be
--- the first thing in the cabal file.
-
--- Initial package description 'intuition' generated by
--- 'cabal init'. For further documentation, see:
---   http://haskell.org/cabal/users-guide/
---
--- The name of the package.
 name:               intuition
-
--- The package version.
--- See the Haskell package versioning policy (PVP) for standards
--- guiding when and how versions should be incremented.
--- https://pvp.haskell.org
--- PVP summary:     +-+------- breaking API changes
---                  | | +----- non-breaking API additions
---                  | | | +--- code changes with no API change
 version:            0.3.0.1
-
--- A short (one-line) description of the package.
 synopsis:           Intuitionistic logic PicoSAT
-
--- A longer description of the package.
--- description:
-
--- URL for the project homepage or repository.
 homepage:           https://github.com/alx/intuition
-
--- The license under which the package is released.
 license:            NONE
-
--- The package author(s).
 author:             Aleksy Schubert
-
--- An email address to which users can send suggestions, bug reports, and patches.
 maintainer:         alx@mimuw.edu.pl
-
--- A copyright notice.
--- copyright:
 category:           Math
 build-type:         Simple
-
--- Extra doc files to be distributed with the package, such as a CHANGELOG or a README.
 extra-doc-files:    CHANGELOG.md
-
--- Extra source files to be distributed with the package, such as examples, or a tutorial module.
--- extra-source-files:
-
 common warnings
     ghc-options: -Wall
 
-executable intuition
-    -- Import common warning flags.
-    import:           warnings
+library
+    exposed-modules:
+      Context
+      Formulas
+      Prover
+    hs-source-dirs: src
+    build-depends:    base >=4.11.3.0,
+                      tptp,
+                      sqlite-simple,
+                      attoparsec,
+                      containers,
+                      text >=1.2.0.0 && <1.5,
+                      timeit
 
-    -- .hs or .lhs file containing the Main module.
+
+executable intuition
+    import:           warnings
     main-is:          Main.hs
 
     -- Modules included in this executable, other than Main.
@@ -72,9 +42,11 @@ executable intuition
                       tptp,
                       sqlite-simple,
                       attoparsec,
-                      containers
+                      containers,
+                      text >=1.2.0.0 && <1.5,
+                      timeit,
+                      intuition
 
-    -- Directories containing source files.
     hs-source-dirs:   app
 
     -- Base language which the package is written in.
